# Properly handle UTF-8 keys and values