`rb_str_new` and `rb_intern` create strings with `ENC_CODERANGE_7BIT` (`Encoding::US_ASCII`), which like it's name implies doesn't allow for anything outside of the ASCII range.

According to the JSON spec, string should be assumed to be UTF-8, so IMHO it's better to just go with that.

Obviously this is a bit slower, but is required for corectness.