@anilmaurya 2020-07-21
@casperisfine How much slower, do you have any benchmark for it ?

I am thinking, if it's making huge performance impact then we can give option to select ASCII or UTF-8, what do you think ?