@casperisfine 2020-07-21
I'll try to benchmark it. But long story short, if the symbol already exist `rb_intern` is able to not allocate a String. It directly lookup the symbol table with the `char *` (I'm simplifying but that's the idea).

So this means each key will always allocated a string, generating more objects to be gced. So it's not really much slower immediately, but generates a bit more GC pressure.

>  we can give option to select ASCII or UTF-8, what do you think ?

It's possible yes. But I would advise to make UTF-8 the default. Super fast parsers are nice, but speed should take a backseat to correcteness.