@casperisfine 2020-07-21
So on my machine, using `file_ips_benchmark.rb`:

current master:

`      FastJsonparser    874.782  (± 2.1%) i/s -      4.450k in   5.089163s`

This branch:

`      FastJsonparser    786.259  (± 2.2%) i/s -      3.978k in   5.061801s`