@casperisfine 2020-07-21
One alternative could be to check in cpp wether the string is ASCII only. It still would be a bit slower, but not as much.