@anilmaurya 2020-07-21
@casperisfine 

> One alternative could be to check in cpp wether the string is ASCII only.

This sounds a good option to me, can we get a benchmark for this option ?
Thank you for suggestion.