@casperisfine 2020-07-21
So I tried that here: https://github.com/casperisfine/fast_jsonparser/commit/c14d0e768d9ac3b425d5ee8bd9f1093ab4822197

master:
```
Warming up --------------------------------------
      FastJsonparser    86.000  i/100ms
                  OJ    56.000  i/100ms
       standard JSON    31.000  i/100ms
Calculating -------------------------------------
      FastJsonparser    850.404  (± 2.6%) i/s -      4.300k in   5.059867s
                  OJ    564.745  (± 1.6%) i/s -      2.856k in   5.058347s
       standard JSON    315.189  (± 1.6%) i/s -      1.581k in   5.017322s
```
this branch:

```
Warming up --------------------------------------
      FastJsonparser    77.000  i/100ms
                  OJ    56.000  i/100ms
       standard JSON    32.000  i/100ms
Calculating -------------------------------------
      FastJsonparser    781.121  (± 1.8%) i/s -      3.927k in   5.029071s
                  OJ    562.968  (± 2.3%) i/s -      2.856k in   5.076014s
       standard JSON    317.943  (± 2.2%) i/s -      1.600k in   5.034690s
```

the ascii check branch:
```
Warming up --------------------------------------
      FastJsonparser    78.000  i/100ms
                  OJ    56.000  i/100ms
       standard JSON    31.000  i/100ms
Calculating -------------------------------------
      FastJsonparser    777.211  (± 2.4%) i/s -      3.900k in   5.021019s
                  OJ    555.251  (± 3.2%) i/s -      2.800k in   5.048275s
       standard JSON    318.336  (± 1.9%) i/s -      1.612k in   5.065718s
```

So at least on that particular benchmark, it's not really much better. I could try to port the SIMD ascii checks, but that would only help for larger strings, so I really don't expect any improvement for objects keys in realistic scenarios.