@casperisfine 2020-07-21
The solution to this might come from simdjson in the future: 

  - https://github.com/simdjson/simdjson/issues/184
  - https://github.com/simdjson/simdjson/issues/264