@casperisfine 2020-07-21
cc @lemire since you commented on this repo, simdjson/simdjson#184 / simdjson/simdjson#264 would be very useful for the performance of the ruby bindings (and possibly other languages?)