@lemire 2020-07-21
@casperisfine  

> The solution to this might come from simdjson in the future 

We are committed to helping the bindings become better. Yet we have to understand what the problems are.


[Issue 184](https://github.com/simdjson/simdjson/issues/184) has to do with how we could have different storage depending on the length of the string. It would have mostly an internal (to simdjson) benefit. I don't think we would expand our API so that you can get "short and normal" strings. That would be a nightmare of usability. It would have to provide very clear performance benefits to the users, and it seems unlikely. I must say that issue 184 is unlikely to ever be implemented, and I think we will drop it. When I first thought about it, I figured that most keys would be very short, but when you look at actual JSON documents, you find plenty of relatively long keys. 


[Issue 264](https://github.com/simdjson/simdjson/issues/264) is more interesting. The request is for a feature where you could ask whether all keys are ASCII. Note that this is only for keys, not for strings in general (as they are less likely to be ASCII). This is justified because, very often, keys are only ASCII. But it is not clear to me why it is best done within simdjson. You can check, for yourself, whether a given key is ASCII. That is easy enough. If people find it convenient, we could certainly provide a helper function which would take a string_view instance and tell you whether it is ASCII.


> I could try to port the SIMD ascii checks, but that would only help for larger strings, so I really don't expect any improvement for objects keys in realistic scenarios.

Given how short keys are, you probably don't need SIMD at all to get great performance and for a problem so simple, compilers can autovectorize (gcc with -O3 and clang with -O2).

As you no doubt realize, checking that a string is ASCII is cheap. It is easy to program.