@lemire 2020-07-21
@casperisfine

I'd like to know more about your expectations. Do you expect the JSON inputs to be all ASCII?