@casperisfine 2020-07-21
@lemire I need to leave my desk soon but I'll try to summarize quickly, I'll extend my answer tomorrow.

In short Ruby have symbols (basically interned strings) and you can build them fast from C extensions with little GC overhead. However that cheap method is only available and valid for pure ASCII strings. If you provide it with with a string containing characters outside the ASCII range, you'll end up with a pretty much broken object.

That's why I opened this PR, to do this in a more correct, but unfortunately slower, way. So I was trying to see if I could conciliate both with a cheap check for ASCII only keys. I did it this way: https://github.com/anilmaurya/fast_jsonparser/commit/c14d0e768d9ac3b425d5ee8bd9f1093ab4822197 but the benchmark didn't improve. Now it's very possible I made a mistake either in the implementation (I'm mostly a C/C++ newbie) or in the benchmark.

I also didn't expect SIMD to be any help for JSON keys.