@lemire 2020-07-21
@casperisfine 
You can probably do a bit better regarding the ASCII check. Check my blog post: [Avoid character-by-character processing when performance matters](https://lemire.me/blog/2020/07/21/avoid-character-by-character-processing-when-performance-matters/).

 It is not going to be a massive gain, but it might push it above your desired threshold.