@casperisfine 2020-07-22
Ok, so I ran the benchmark again, and honestly the difference between `rb_intern2` and `rb_intern_str(rb_utf8_str_new())` isn't that big at all. I must have done something wrong in my benchmark, or simply had bad variance.

Thanks for the advices!